# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 黄冠AI Academy API
  version: 2.0.0
  description: 学生教育培训平台 API（/api/v1），含错题分析、预约、学习总结与练习模块。
servers:
  - url: /api/v1
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yy_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
paths:
...
  /wrong-questions:
    get:
      summary: 错题列表（家庭）
      security:
        - sessionCookie: []
      parameters:
        - name: subject
          in: query
          schema:
            type: string
            enum: [MATH, ENGLISH]
        - name: mastered
          in: query
          schema:
            type: string
            enum: ["true", "false"]
        - name: knowledgePoint
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 错题列表
    post:
      summary: 上传错题图片
      security:
        - sessionCookie: []
        - csrfHeader: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [subject, image]
              properties:
                subject:
                  type: string
                  enum: [MATH, ENGLISH]
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: 上传成功，返回 PROCESSING
...